+++ src/shinestacker/algorithms/align.py
@@ -169,7 +169,7 @@ def align_images(img_1, img_0, feature_config=None, matching_config=None, alignm
     subsample = alignment_config['subsample']
     if subsample == 0:
         img_res = (float(h0) / 1000) * (float(w0) / 1000)
-        target_res = constants.DEFALUT_ALIGN_RES_TARGET_MPX
+        target_res = constants.DEFAULT_ALIGN_RES_TARGET_MPX
         subsample = int(1 + math.floor(img_res / target_res))
     fast_subsampling = alignment_config['fast_subsampling']
     min_good_matches = alignment_config['min_good_matches']