+++ src/shinestacker/gui/project_converter.py
@@ -9,8 +9,7 @@ from .. algorithms.vignetting import Vignetting
 from .. algorithms.align import AlignFrames
 from .. algorithms.balance import BalanceFrames
 from .. algorithms.stack import FocusStack, FocusStackBunch
-from .. algorithms.pyramid import PyramidStack
-from .. algorithms.pyramid_tiles import PyramidTilesStack
+from .. algorithms.pyramid_auto import PyramidAutoStack
 from .. algorithms.depth_map import DepthMapStack
 from .. algorithms.multilayer import MultiLayer
 from .project_model import Project, ActionConfig
@@ -107,11 +106,7 @@ class ProjectConverter:
             if stacker == constants.STACK_ALGO_PYRAMID:
                 algo_dict, module_dict = self.filter_dict_keys(
                     action_config.params, 'pyramid_')
-                stack_algo = PyramidStack(**algo_dict)
-            elif stacker == constants.STACK_ALGO_PYRAMID_TILES:
-                algo_dict, module_dict = self.filter_dict_keys(
-                    action_config.params, 'tiles_pyramid_')
-                stack_algo = PyramidTilesStack(**algo_dict)
+                stack_algo = PyramidAutoStack(**algo_dict)
             elif stacker == constants.STACK_ALGO_DEPTH_MAP:
                 algo_dict, module_dict = self.filter_dict_keys(
                     action_config.params, 'depthmap_')
@@ -119,7 +114,6 @@ class ProjectConverter:
             else:
                 raise InvalidOptionError('stacker', stacker, f"valid options are: "
                                          f"{constants.STACK_ALGO_PYRAMID}, "
-                                         f"{constants.STACK_ALGO_PYRAMID_TILES}, "
                                          f"{constants.STACK_ALGO_DEPTH_MAP}")
             if action_config.type_name == constants.ACTION_FOCUSSTACK:
                 return FocusStack(**module_dict, stack_algo=stack_algo)